Exclude 'Logs' folder from being deleted during deployment of WebAPI
When files are deleted the application might still be running and log files might still be locked, so to avoid errors let's exclude log files from being deleted during deployment.
Also, it is a good idea to not delete log files during deployment as they might be useful for future analysis.