Suppress the context menu
If context menu is up every click event (with any either left or right mouse buttons) has shifted x coordinate (-1 less than actual position of the mouse). Because of that we need to suppress the menu so it does not cause side-effects.
Also, it is just not required on the canvas as right-click event should provide another functionality.
Unused 'event' parameter was added to 'onCanvasMouseUp' as it might be useful later (for logging) and to avoid further unnecessary changes to html file.