Implemented test for checking validation logic for 'nick' field
The major bug was in setting the value for the control (see changes to the playerNick setter method).

Validation of 'register' button is implemented in 2 ways: via Harness and via 'native elements'. The reason that both are kept is that it was hard to identify the proper way and I was experimenting with both.
In the end, I would like to keep a working example of each of them.

https://stackoverflow.com/questions/69323657/reactive-forms-why-2-way-biding-works-differently-under-jasmine-karma
https://stackblitz.com/edit/stackoverflow-q-53122846-szzgrz?file=main.ts